♻️ refactor: js => ts 변경 작업 및 에러 메시지 추가
