+++ src/index.js
@@ -1,5 +1,5 @@
 import { App } from "./App";
-import Css from "./css/style.css";
+import "./css/style.css";
 import "./assets/add-button.png";
 import "./assets/category-asian.png";
 import "./assets/category-korean.png";
